+++ packages/soda-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-react
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.35.0
+
 ## 0.0.6
 
 ### Patch Changes